Typescript refactoring of matcher. Moved c8ypact related code in c8ypact.ts
