+++ lib/commands/c8yclient.js
@@ -8,7 +8,6 @@ import {
   tenantFromBasicAuth,
   throwError,
 } from "./utils";
-import { C8yDefaultPactMatcher } from "./../pacts/matcher";
 
 const { _ } = Cypress;
 
@@ -20,17 +19,6 @@ const defaultOptions = {
   skipClientAuthenication: false,
 };
 
-Cypress.c8ypact = {
-  currentPactIdentifier: pactIdentifier,
-  currentPacts,
-  currentPactFilename,
-  currentNextPact: getNextPact,
-  isRecordingEnabled,
-  matcher: new C8yDefaultPactMatcher(),
-  failOnMissingPacts: false,
-};
-
-const logTasks = false;
 let logOnce = true;
 
 window.fetchStub = window.fetch;
@@ -279,7 +267,7 @@ function preprocessResponseObject(r, save = true) {
   }
   response.method = (r.res && r.res.method) || r.method || "GET";
   if (Cypress.c8ypact.isRecordingEnabled() && save) {
-    savePact(response);
+    Cypress.c8ypact.savePact(response);
   }
   return response;
 }
@@ -374,7 +362,7 @@ function run(client, fns, prevSubject, options, baseUrl) {
 
     if (!Cypress.c8ypact.isRecordingEnabled()) {
       for (const r of _.isArray(response) ? response : [response]) {
-        getNextPact().then((pact) => {
+        Cypress.c8ypact.currentNextPact().then((pact) => {
           if (!pact && Cypress.c8ypact.failOnMissingPacts) {
             throwError(`Pact ${Cypress.c8ypact.currentPactIdentifier()} not found and Cypress.c8ypact.failOnMissingPacts is enabled.`);
           }
@@ -456,62 +444,3 @@ function isArrayOfFunctions(functions) {
   if (!functions || !_.isArray(functions)) return false;
   return _.isEmpty(functions.filter((f) => !_.isFunction(f)));
 }
-
-function pactIdentifier() {
-  let key = Cypress.currentTest?.titlePath?.join("--");
-  if (key == null) {
-    key = Cypress.spec?.relative?.split("/").slice(-2).join("--");
-  }
-  return Cypress.config().c8ypact || key.replace(/ /g, "_");
-}
-
-function isRecordingEnabled() {
-  return Cypress.env("C8Y_PACT_MODE") === "recording";
-}
-
-function savePact(response) {
-  const pact = Cypress.c8ypact.currentPactIdentifier();
-  if (pact) {
-    const folder = Cypress.config().fixturesFolder;
-    cy.task(
-      "c8ypact:save",
-      {
-        pact,
-        response,
-        folder,
-      },
-      { log: logTasks }
-    );
-  }
-}
-
-function currentPacts() {
-  return cy.task("c8ypact:get", Cypress.c8ypact.currentPactIdentifier(), {
-    log: logTasks,
-  });
-}
-
-function currentPactFilename() {
-  const pactId = Cypress.c8ypact.currentPactIdentifier();
-  return `${Cypress.config().fixturesFolder}/c8ypact/${pactId}.json`;
-}
-
-function getNextPact() {
-  return cy.task("c8ypact:next", Cypress.c8ypact.currentPactIdentifier(), {
-    log: logTasks,
-  });
-}
-
-before(() => {
-  if (!Cypress.c8ypact.isRecordingEnabled()) {
-    cy.task("c8ypact:load", Cypress.config().fixturesFolder, { log: logTasks });
-  }
-});
-
-beforeEach(() => {
-  if (Cypress.c8ypact.isRecordingEnabled()) {
-    cy.task("c8ypact:remove", Cypress.c8ypact.currentPactIdentifier(), {
-      log: logTasks,
-    });
-  }
-});
