+++ lib/commands/index.d.ts
@@ -2,4 +2,5 @@ import './dates'
 import './general'
 import './login'
 import './c8yclient'
+import './c8ypact'
 import "./administration";
