+++ lib/commands/index.js
@@ -2,4 +2,5 @@ import "./general";
 import "./dates";
 import "./login";
 import "./c8yclient";
+import "./c8ypact";
 import "./administration";
