+++ lib/commands/utils.js
@@ -211,8 +211,3 @@ export function throwError(message) {
   throw newErr;
 }
 
-export function throwPactError(message) {
-  const newErr = new Error(message);
-  newErr.name = "C8yPactError";
-  throw newErr;
-}
